requirements.txt
streamlit
yfinance
pandas
matplotlib
ta
openai
requests
beautifulsoup4
lxml
polygon-api-client